+++ core/CMakeLists.txt
@@ -26,8 +26,9 @@ set(INC_FILES
     include/scipp/core/element/logical.h
     include/scipp/core/element/math.h
     include/scipp/core/element/rebin.h
+    include/scipp/core/element/reduction.h
+    include/scipp/core/element/special_values.h
     include/scipp/core/element/trigonometry.h
-    include/scipp/core/element/unary_operations.h
     include/scipp/core/element/util.h
 )
