+++ core/include/scipp/core/element/special_values.h
