+++ core/test/CMakeLists.txt
@@ -15,8 +15,8 @@ add_executable(
   element_logical_test.cpp
   element_math_test.cpp
   element_reduction_test.cpp
+  element_special_values_test.cpp
   element_trigonometry_test.cpp
-  element_unary_operations_test.cpp
   element_util_test.cpp
   value_and_variance_test.cpp
   view_index_test.cpp
